neurons/speech_to_text/miners/librispeech/requirements.txt
torchaudio
transformers[speech, sentencepiece]
soundfile